+++ examples/inverse/plot_mne_point_spread_function.py
@@ -13,9 +13,8 @@ across the cortical surface.
 #
 # License: BSD (3-clause)
 
-from mayavi import mlab
-
 import mne
+from mne.viz import create_3d_figure
 from mne.datasets import sample
 from mne.minimum_norm import read_inverse_operator, point_spread_function
 
@@ -66,12 +65,12 @@ stc_psf_meg, _ = point_spread_function(
 time_label = "EEGMEG %d"
 brain_eegmeg = stc_psf_eegmeg.plot(hemi='rh', subjects_dir=subjects_dir,
                                    time_label=time_label,
-                                   figure=mlab.figure(size=(500, 500)))
+                                   figure=create_3d_figure(size=(500, 500)))
 
 time_label = "MEG %d"
 brain_meg = stc_psf_meg.plot(hemi='rh', subjects_dir=subjects_dir,
                              time_label=time_label,
-                             figure=mlab.figure(size=(500, 500)))
+                             figure=create_3d_figure(size=(500, 500)))
 
 # The PSF is centred around the right auditory cortex label,
 # but clearly extends beyond it.