+++ mne/io/brainvision/tests/test_brainvision.py
@@ -6,6 +6,7 @@
 
 import os.path as op
 from os import unlink
+import warnings
 import shutil
 
 import numpy as np
@@ -15,6 +16,7 @@ import pytest
 from tempfile import NamedTemporaryFile
 
 from mne.utils import _TempDir, run_tests_if_main
+from mne.utils.numerics import object_diff
 from mne import pick_types, read_annotations, concatenate_raws
 from mne.io.constants import FIFF
 from mne.io import read_raw_fif, read_raw_brainvision
@@ -60,6 +62,33 @@ vhdr_bad_date = op.join(data_dir, 'test_bad_date.vhdr')
 montage = op.join(data_dir, 'test.hpts')
 eeg_bin = op.join(data_dir, 'test_bin_raw.fif')
 eog = ['HL', 'HR', 'Vb']
+
+
+@pytest.mark.parametrize('montage', [montage, 'biosemi32'])
+def test_same_behaviour_in_init_and_set_montage(montage):
+    """Test that __init__ and set_montage lead to equal results."""
+    with pytest.warns(RuntimeWarning) as init_warns:
+        warnings.warn('dummy', RuntimeWarning)
+        raw_montage = read_raw_brainvision(vhdr_path, montage=montage)
+
+    raw_none = read_raw_brainvision(vhdr_path, montage=None)
+    assert raw_none.info['dig'] is None
+
+    with pytest.warns(RuntimeWarning) as set_montage_warns:
+        warnings.warn('dummy', RuntimeWarning)
+        raw_none.set_montage(montage)
+
+    # Assert equal objects
+    for key in ['chs', 'dig']:
+        diff = object_diff(raw_none.info[key], raw_montage.info[key])
+        assert diff == ''
+
+    # Assert equal warnings
+    assert len(init_warns) == len(set_montage_warns)
+    for ii in range(len(init_warns)):
+        msg_a = init_warns[ii].message.args[0]
+        msg_b = set_montage_warns[ii].message.args[0]
+        assert msg_a == msg_b
 
 
 def test_orig_units(recwarn):