+++ mne/tests/test_docstring_parameters.py
@@ -24,7 +24,6 @@ public_modules = [
     'mne.datasets',
     'mne.datasets.brainstorm',
     'mne.datasets.hf_sef',
-    'mne.datasets.megsim',
     'mne.datasets.sample',
     'mne.decoding',
     'mne.dipole',
@@ -216,7 +215,6 @@ BaseEstimator
 ContainsMixin
 CrossSpectralDensity
 FilterMixin
-HilbertMixin
 GeneralizationAcrossTime
 RawFIF
 TimeMixin