+++ mne/viz/tests/test_3d.py
@@ -24,7 +24,8 @@ from mne.io.pick import pick_info
 from mne.io.constants import FIFF
 from mne.viz import (plot_sparse_source_estimates, plot_source_estimates,
                      snapshot_brain_montage, plot_head_positions,
-                     plot_alignment, plot_volume_source_estimates)
+                     plot_alignment, plot_volume_source_estimates,
+                     plot_sensors_connectivity)
 from mne.viz.utils import _fake_click
 from mne.utils import (requires_mayavi, requires_pysurfer, run_tests_if_main,
                        _import_mlab, requires_nibabel, check_version,
@@ -250,13 +251,16 @@ def test_plot_alignment(tmpdir, renderer):
                    surfaces=['head', 'inner_skull'], bem=bem_surfs)
     # single-layer BEM can still plot head surface
     assert bem_surfs[-1]['id'] == FIFF.FIFFV_BEM_SURF_ID_BRAIN
-    with catch_logging() as log:
-        plot_alignment(info, trans_fname, subject='sample',
-                       meg=True, subjects_dir=subjects_dir,
-                       surfaces=['head', 'inner_skull'], bem=bem_surfs[-1:],
-                       verbose=True)
-    log = log.getvalue()
-    assert 'not find the surface for head in the provided BEM model' in log
+    bem_sol_homog = read_bem_solution(op.join(subjects_dir, 'sample', 'bem',
+                                              'sample-1280-bem-sol.fif'))
+    for use_bem in (bem_surfs[-1:], bem_sol_homog):
+        with catch_logging() as log:
+            plot_alignment(info, trans_fname, subject='sample',
+                           meg=True, subjects_dir=subjects_dir,
+                           surfaces=['head', 'inner_skull'], bem=use_bem,
+                           verbose=True)
+        log = log.getvalue()
+        assert 'not find the surface for head in the provided BEM model' in log
     # sphere model
     sphere = make_sphere_model('auto', 'auto', evoked.info)
     src = setup_volume_source_space(sphere=sphere)
@@ -506,5 +510,30 @@ def test_plot_vec_source_estimates():
         stc.plot('sample', subjects_dir=subjects_dir,
                  clim=dict(pos_lims=[1, 2, 3]))
 
+
+@testing.requires_testing_data
+@requires_mayavi
+def test_plot_sensors_connectivity():
+    """Test plotting of sensors connectivity."""
+    from mne import io, pick_types
+
+    data_path = data_dir
+    raw_fname = op.join(data_path, 'MEG', 'sample',
+                        'sample_audvis_trunc_raw.fif')
+
+    raw = io.read_raw_fif(raw_fname)
+    picks = pick_types(raw.info, meg='grad', eeg=False, stim=False,
+                       eog=True, exclude='bads')
+    n_channels = len(picks)
+    con = np.random.RandomState(42).randn(n_channels, n_channels)
+    info = raw.info
+    with pytest.raises(TypeError):
+        plot_sensors_connectivity(info='foo', con=con,
+                                  picks=picks)
+    with pytest.raises(ValueError):
+        plot_sensors_connectivity(info=info, con=con[::2, ::2],
+                                  picks=picks)
+
+    plot_sensors_connectivity(info=info, con=con, picks=picks)
 
 run_tests_if_main()